Move staff tenant auth tests to  their own file
